requirements.txt
requests
types-requests
beautifulsoup4
python-dotenv
discord.py
environ-config
Pillow
ImageHash
